+++ integration-tests/ssr/test-output.js
@@ -42,6 +42,15 @@
       )
     )
 
+    // Fetch once to trigger re-compilation.
+    await fetch(`${devSiteBasePath}/${path}`)
+
+    // Then wait for 6 seconds to ensure it's ready to go.
+    // Otherwise, tests are flaky depending on the speed of the testing machine.
+    await new Promise(resolve => {
+      setTimeout(() => resolve(), 6000)
+    })
+
     let devStatus = 200
     const rawDevHtml = await fetch(`${devSiteBasePath}/${path}`).then(res => {
       devStatus = res.status