+++ packages/gatsby/src/commands/build.ts
@@ -4,7 +4,7 @@ import signalExit from "signal-exit"
 import fs from "fs-extra"
 import telemetry from "gatsby-telemetry"
 
-import { buildHTML } from "./build-html"
+import { doBuildPages, buildRenderer, deleteRenderer } from "./build-html"
 import { buildProductionBundle } from "./build-javascript"
 import { bootstrap } from "../bootstrap"
 import apiRunnerNode from "../utils/api-runner-node"
@@ -57,6 +57,7 @@ interface IBuildArgs extends IProgram {
   profile: boolean
   graphqlTracing: boolean
   openTracingConfigFile: string
+  keepPageRenderer: boolean
 }
 
 module.exports = async function build(program: IBuildArgs): Promise<void> {
@@ -231,6 +232,20 @@ module.exports = async function build(program: IBuildArgs): Promise<void> {
     }
   }
 
+  const buildSSRBundleActivityProgress = report.activityTimer(
+    `Building HTML renderer`,
+    { parentSpan: buildSpan }
+  )
+  buildSSRBundleActivityProgress.start()
+  let pageRenderer: string
+  try {
+    pageRenderer = await buildRenderer(program, Stage.BuildHTML, buildSpan)
+  } catch (err) {
+    buildActivityTimer.panic(structureWebpackErrors(Stage.BuildHTML, err))
+  } finally {
+    buildSSRBundleActivityProgress.end()
+  }
+
   const buildHTMLActivityProgress = report.createProgress(
     `Building static HTML for pages`,
     pagePaths.length,
@@ -241,13 +256,12 @@ module.exports = async function build(program: IBuildArgs): Promise<void> {
   )
   buildHTMLActivityProgress.start()
   try {
-    await buildHTML({
-      program,
-      stage: Stage.BuildHTML,
+    await doBuildPages(
+      pageRenderer,
       pagePaths,
-      activity: buildHTMLActivityProgress,
-      workerPool,
-    })
+      buildHTMLActivityProgress,
+      workerPool
+    )
   } catch (err) {
     let id = `95313` // TODO: verify error IDs exist
     const context = {
@@ -270,6 +284,14 @@ module.exports = async function build(program: IBuildArgs): Promise<void> {
     })
   }
   buildHTMLActivityProgress.end()
+
+  if (!program.keepPageRenderer) {
+    try {
+      await deleteRenderer(pageRenderer)
+    } catch (err) {
+      // pass through
+    }
+  }
 
   let deletedPageKeys: Array<string> = []
   if (process.env.GATSBY_EXPERIMENTAL_PAGE_BUILD_ON_DATA_CHANGES) {