+++ packages/gatsby/src/utils/flags.ts
@@ -1,4 +1,3 @@
-import sampleSiteForExperiment from "./sample-site-for-experiment"
 import _ from "lodash"
 import semver from "semver"
 
@@ -29,14 +28,17 @@ export const satisfiesSemvers = (
   return result
 }
 
-export type fitnessEnum = true | false | "OPT_IN"
+export type fitnessEnum = true | false | "OPT_IN" | "LOCKED_IN"
 
 export interface IFlag {
   name: string
   env: string
   description: string
   command: executingCommand
-  telemetryId: string
+  /**
+   * Use string identifier to track enabled flag or false to disable any tracking (useful when flag becomes new defaults)
+   */
+  telemetryId: string | false
   // Heuristics for deciding if a flag is experimental:
   // - there are known bugs most people will encounter and that block being
   // able to use Gatsby normally
@@ -47,13 +49,20 @@ export interface IFlag {
   // resolved and ~50+ people have tested it, experimental should be set to
   // false.
   experimental: boolean
-  // Generally just return true.
-  //
-  // False means it'll be disabled despite the user setting it true e.g.
-  // it just won't work e.g. it doesn't have new enough version for something.
-  //
-  // OPT_IN means the gatsby will enable the flag (unless the user explicitly
-  // disablees it.
+  /**
+   * True means conditions for the feature are met and can be opted in by user.
+   *
+   * False means it'll be disabled despite the user setting it true e.g.
+   * it just won't work e.g. it doesn't have new enough version for something.
+   *
+   * OPT_IN means the gatsby will enable the flag (unless the user explicitly
+   * disables it.
+   *
+   * LOCKED_IN means that feature is enabled always (unless `noCI` condition is met).
+   * This is mostly to provide more meaningful terminal messages instead of removing
+   * flag from the flag list when users has the flag set in configuration
+   * (avoids showing unknown flag message and shows "no longer needed" message).
+   */
   testFitness: (flag: IFlag) => fitnessEnum
   includedFlags?: Array<string>
   umbrellaIssue?: string
@@ -70,8 +79,6 @@ const activeFlags: Array<IFlag> = [
     description: `Enable all experiments aimed at improving develop server start time`,
     includedFlags: [
       `DEV_SSR`,
-      `QUERY_ON_DEMAND`,
-      `LAZY_IMAGES`,
       `PRESERVE_FILE_DOWNLOAD_CACHE`,
       `PRESERVE_WEBPACK_CACHE`,
     ],
@@ -91,70 +98,33 @@ const activeFlags: Array<IFlag> = [
     name: `QUERY_ON_DEMAND`,
     env: `GATSBY_EXPERIMENTAL_QUERY_ON_DEMAND`,
     command: `develop`,
-    telemetryId: `QueryOnDemand`,
+    telemetryId: false,
     experimental: false,
     description: `Only run queries when needed instead of running all queries upfront. Speeds starting the develop server.`,
     umbrellaIssue: `https://gatsby.dev/query-on-demand-feedback`,
     noCI: true,
-    testFitness: (): fitnessEnum => {
-      // Take a 10% of slice of users.
-      if (sampleSiteForExperiment(`QUERY_ON_DEMAND`, 10)) {
-        let isPluginSharpNewEnoughOrNotInstalled = false
-        try {
-          // Try requiring plugin-sharp so we know if it's installed or not.
-          // If it does, we also check if it's new enough.
-          // eslint-disable-next-line
-          require.resolve(`gatsby-plugin-sharp`)
-          const semverConstraints = {
-            // Because of this, this flag will never show up
-            "gatsby-plugin-sharp": `>=2.10.0`,
-          }
-          if (satisfiesSemvers(semverConstraints)) {
-            isPluginSharpNewEnoughOrNotInstalled = true
-          }
-        } catch (e) {
-          if (e.code === `MODULE_NOT_FOUND`) {
-            isPluginSharpNewEnoughOrNotInstalled = true
-          }
-        }
-
-        if (isPluginSharpNewEnoughOrNotInstalled) {
-          return `OPT_IN`
-        } else {
-          // Don't opt them in but they can still manually enable.
-          return true
-        }
-      } else {
-        // Don't opt them in but they can still manually enable.
-        return true
-      }
-    },
+    testFitness: (): fitnessEnum => `LOCKED_IN`,
   },
   {
     name: `LAZY_IMAGES`,
     env: `GATSBY_EXPERIMENTAL_LAZY_IMAGES`,
     command: `develop`,
-    telemetryId: `LazyImageProcessing`,
+    telemetryId: false,
     experimental: false,
     description: `Don't process images during development until they're requested from the browser. Speeds starting the develop server. Requires gatsby-plugin-sharp@2.10.0 or above.`,
     umbrellaIssue: `https://gatsby.dev/lazy-images-feedback`,
     noCI: true,
     testFitness: (): fitnessEnum => {
-      // Take a 10% of slice of users.
-      if (sampleSiteForExperiment(`QUERY_ON_DEMAND`, 10)) {
-        const semverConstraints = {
-          // Because of this, this flag will never show up
-          "gatsby-plugin-sharp": `>=2.10.0`,
-        }
-        if (satisfiesSemvers(semverConstraints)) {
-          return `OPT_IN`
-        } else {
-          // gatsby-plugi-sharp is either not installed or not new enough so
-          // just disable — it won't work anyways.
-          return false
-        }
+      const semverConstraints = {
+        // Because of this, this flag will never show up
+        "gatsby-plugin-sharp": `>=2.10.0`,
+      }
+      if (satisfiesSemvers(semverConstraints)) {
+        return `LOCKED_IN`
       } else {
-        return true
+        // gatsby-plugin-sharp is either not installed or not new enough so
+        // just disable — it won't work anyways.
+        return false
       }
     },
   },