+++ packages/gatsby/src/utils/webpack-error-utils.ts
@@ -3,15 +3,12 @@ import { Stats } from "webpack"
 import { IMatch } from "../types"
 import { Stage as StageEnum } from "../commands/types"
 
-const stageCodeToReadableLabel = {
-  BuildJavascript: `Generating JavaScript bundles`,
-  BuildHTML: `Generating SSR bundle`,
-  DevelopHTML: `Generating development SSR bundle`,
-  Develop: `Generating development JavaScript bundle`,
-} as const
-
-type Stage = keyof typeof StageEnum
-type StageLabel = typeof stageCodeToReadableLabel[Stage]
+const stageCodeToReadableLabel: Record<StageEnum, string> = {
+  [StageEnum.BuildJavascript]: `Generating JavaScript bundles`,
+  [StageEnum.BuildHTML]: `Generating SSR bundle`,
+  [StageEnum.DevelopHTML]: `Generating development SSR bundle`,
+  [StageEnum.Develop]: `Generating development JavaScript bundle`,
+}
 
 interface ITransformedWebpackError {
   id: string
@@ -21,7 +18,7 @@ interface ITransformedWebpackError {
   }
   context: {
     stage: StageEnum
-    stageLabel: StageLabel
+    stageLabel: string
     sourceMessage?: string
     [key: string]: unknown
   }